ncaab_h1_model/requirements.txt
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
python-dotenv>=1.0.0
